requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
opencv-python==4.8.1.78
mediapipe==0.10.8
deepface==0.0.75
pillow==10.1.0
scikit-image==0.22.0
openai-whisper==20231117
librosa==0.10.0
soundfile==0.12.1
transformers==4.35.2
sentence-transformers==2.2.2
numpy==1.26.2
scipy==1.11.4
torch==2.1.1
torchaudio==2.1.1
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
openai==1.3.3
anthropic==0.7.1
reportlab==4.0.9
pypdf==3.17.1
python-jose==3.3.0
passlib==1.7.4
